querymethods should operate on the original caller.
Since Manager.__getattr__ was passing straight through to the QuerySet
rather than operating on the Manager object itself, get_or_create broke
on relations. The querymethod could see a QMethodQuerySet but not the
original RelatedManager instance that get_or_create was called on,
resulting in IntegrityErrors.

I refactored the __getattr__ mechanism into a mixin which is now added
to both the QuerySet and the Manager; this means QuerySets aren't
instantiated for every missing attribute lookup.